[ #23 | Refactor ] user, auth 타입 분리리
