+++ src/types/auth.type.ts
@@ -1,5 +1,28 @@
 import { User } from '@/types/user.type';
 
+// 회원가입 요청 타입
+export interface SignupRequest {
+  email: string;
+  password: string;
+  username: string;
+}
+
+// 로그인 요청 타입
+export interface LoginRequest {
+  email: string;
+  password: string;
+}
+
+// 로그인 응답 타입
+export interface LoginResponse {
+  status: string;
+  message: string;
+  data: {
+    accessToken: string;
+  };
+  code: number;
+}
+
 // 토큰 정보 타입
 export interface AuthState {
   token: string | null;
